+++ hera_cal/tests/test_frf.py
@@ -247,19 +247,19 @@ class Test_FRFilter(object):
         assert args.flag_output is None
         assert args.filetype == "uvh5"
 
-    def test_run_fr_filter(self):
+    def test_run_tophat_frfilter(self):
         fname = os.path.join(DATA_PATH, "zen.2458043.12552.xx.HH.uvORA")
         k = (24, 25, 'ee')
-        frfil = frf.FRFilter(fname, filetype='miriad')
+        tfrfil = tfrf.FRFilter(fname, filetype='miriad')
         frfil.read(bls=[k])
         bl = np.linalg.norm(frfil.antpos[24] - frfil.antpos[25]) / constants.c * 1e9
         sdf = (frfil.freqs[1] - frfil.freqs[0]) / 1e9
 
-        frfil.run_fr_filter(to_filter=frfil.data.keys(), tol=1e-2)
+        frfil.run_tophat_frfilter(to_filter=frfil.data.keys(), tol=1e-2, output_prefix='frfiltered')
         for k in frfil.data.keys():
-            assert frfil.clean_resid[k].shape == (60, 64)
-            assert frfil.clean_model[k].shape == (60, 64)
-            assert k in frfil.clean_info
+            assert frfil.frfiltered_resid[k].shape == (60, 64)
+            assert frfil.frfiltered_model[k].shape == (60, 64)
+            assert k in frfil.frfiltered_info
 
         # test skip_wgt imposition of flags
         fname = os.path.join(DATA_PATH, "zen.2458043.12552.xx.HH.uvORA")
@@ -272,13 +272,13 @@ class Test_FRFilter(object):
                 frfil.avg_red_baseline_vectors()
             wgts = {k: np.ones_like(frfil.flags[k], dtype=np.float)}
             wgts[k][:, 0] = 0.0
-            frfil.run_fr_filter(to_filter=[k], weight_dict=wgts, tol=1e-5, window='blackman-harris', skip_wgt=0.1, maxiter=100)
+            frfil.run_tophat_frfilter(to_filter=[k], weight_dict=wgts, tol=1e-5, window='blackman-harris', skip_wgt=0.1, maxiter=100)
             assert frfil.clean_info[k]['status']['axis_0'][0] == 'skipped'
             np.testing.assert_array_equal(frfil.clean_flags[k][:, 0], np.ones_like(frfil.flags[k][:, 0]))
             np.testing.assert_array_equal(frfil.clean_model[k][:, 0], np.zeros_like(frfil.clean_resid[k][:, 0]))
             np.testing.assert_array_equal(frfil.clean_resid[k][:, 0], np.zeros_like(frfil.clean_resid[k][:, 0]))
 
-    def test_load_fr_filter_and_write_baseline_list(self, tmpdir):
+    def test_load_tophat_frfilter_and_write_baseline_list(self, tmpdir):
         tmp_path = tmpdir.strpath
         uvh5 = [os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.OCR_53x_54x_only.first.uvh5"),
                 os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.OCR_53x_54x_only.second.uvh5")]
@@ -291,22 +291,22 @@ class Test_FRFilter(object):
             shutil.rmtree(cdir)
         os.mkdir(cdir)
         for avg_bl in [True, False]:
-            frf.load_fr_filter_and_write_baseline_list(datafile_list=uvh5, baseline_list=[(53, 54, 'ee')],
-                                                       calfile_list=cals, spw_range=[100, 200], cache_dir=cdir,
-                                                       read_cache=True, write_cache=True, avg_red_bllens=avg_bl,
-                                                       res_outfilename=outfilename, clobber=True,
-                                                       mode='dpss_leastsq')
+            frf.load_tophat_frfilter_and_write(datafile_list=uvh5, baseline_list=[(53, 54, 'ee')],
+                                                calfile_list=cals, spw_range=[100, 200], cache_dir=cdir,
+                                                read_cache=True, write_cache=True, avg_red_bllens=avg_bl,
+                                                res_outfilename=outfilename, clobber=True,
+                                                mode='dayenu')
             hd = io.HERAData(outfilename)
             d, f, n = hd.read()
             assert len(list(d.keys())) == 1
             assert d[(53, 54, 'ee')].shape[1] == 100
             assert d[(53, 54, 'ee')].shape[0] == 60
             # now do no spw range and no cal files just to cover those lines.
-            frf.load_fr_filter_and_write_baseline_list(datafile_list=uvh5, baseline_list=[(53, 54, 'ee')],
-                                                       cache_dir=cdir,
-                                                       read_cache=True, write_cache=True, avg_red_bllens=avg_bl,
-                                                       res_outfilename=outfilename, clobber=True,
-                                                       mode='dpss_leastsq')
+            frf.load_tophat_frfilter_and_write(datafile_list=uvh5, baseline_list=[(53, 54, 'ee')],
+                                                cache_dir=cdir,
+                                                read_cache=True, write_cache=True, avg_red_bllens=avg_bl,
+                                                res_outfilename=outfilename, clobber=True,
+                                                mode='dayenu')
             hd = io.HERAData(outfilename)
             d, f, n = hd.read()
             assert len(list(d.keys())) == 1
@@ -336,10 +336,11 @@ class Test_FRFilter(object):
         time_thresh = 2. / hd.Ntimes
         for blnum, bl in enumerate(flags.keys()):
             outfilename = os.path.join(tmp_path, 'bl_chunk_%d.h5' % blnum)
-            frf.load_fr_filter_and_write_baseline_list(datafile_list=[input_file], res_outfilename=outfilename,
-                                                       tol=1e-4, baseline_list=[bl],
-                                                       cache_dir=cdir,
-                                                       factorize_flags=True, time_thresh=time_thresh, clobber=True)
+            frf.load_tophat_frfilter_and_write(datafile_list=[input_file], res_outfilename=outfilename,
+                                                tol=1e-4, baseline_list=[bl],
+                                                cache_dir=cdir,
+                                                factorize_flags=True,
+                                                time_thresh=time_thresh, clobber=True)
         # now load all of the outputs in
         output_files = glob.glob(tmp_path + '/bl_chunk_*.h5')
         hd = io.HERAData(output_files)
@@ -359,21 +360,21 @@ class Test_FRFilter(object):
         uvf.flag_array[:] = False
         flagfile = os.path.join(tmp_path, 'test_flag.h5')
         uvf.write(flagfile, clobber=True)
-        frf.load_fr_filter_and_write_baseline_list(datafile_list=[input_file], res_outfilename=outfilename,
-                                                   tol=1e-4, baseline_list=[bl[:2]],
-                                                   clobber=True, mode='dpss_leastsq',
-                                                   external_flags=flagfile, overwrite_flags=True)
+        frf.load_tophat_frfilter_and_write(datafile_list=[input_file], res_outfilename=outfilename,
+                                            tol=1e-4, baseline_list=[bl[:2]],
+                                            clobber=True, mode='dayenu',
+                                            external_flags=flagfile, overwrite_flags=True)
         # test that all flags are False
         hd = io.HERAData(outfilename)
         d, f, n = hd.read()
         for k in f:
             assert np.all(~f[k])
         # now do the external yaml
-        frf.load_fr_filter_and_write_baseline_list(datafile_list=[input_file], res_outfilename=outfilename,
-                                                   tol=1e-4, baseline_list=[bl[:2]],
-                                                   clobber=True, mode='dayenu',
-                                                   external_flags=flagfile, overwrite_flags=True,
-                                                   flag_yaml=flag_yaml)
+        frf.load_tophat_frfilter_and_write(datafile_list=[input_file], res_outfilename=outfilename,
+                                            tol=1e-4, baseline_list=[bl[:2]],
+                                            clobber=True, mode='dayenu',
+                                            external_flags=flagfile, overwrite_flags=True,
+                                            flag_yaml=flag_yaml)
         # test that all flags are af yaml flags
         hd = io.HERAData(outfilename)
         d, f, n = hd.read()
@@ -385,11 +386,11 @@ class Test_FRFilter(object):
         os.remove(outfilename)
         shutil.rmtree(cdir)
 
-    def test_load_fr_filter_and_write(self, tmpdir):
+    def test_load_tophat_frfilter_and_write(self, tmpdir):
         tmp_path = tmpdir.strpath
         uvh5 = os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.OCR_53x_54x_only.uvh5")
         outfilename = os.path.join(tmp_path, 'temp.h5')
-        frf.load_fr_filter_and_write(uvh5, res_outfilename=outfilename, tol=1e-4, clobber=True, Nbls_per_load=1)
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename, tol=1e-4, clobber=True, Nbls_per_load=1)
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for bl in d:
@@ -397,16 +398,19 @@ class Test_FRFilter(object):
 
         frfil = frf.FRFilter(uvh5, filetype='uvh5')
         frfil.read(bls=[(53, 54, 'ee')])
-        frfil.run_fr_filter(to_filter=[(53, 54, 'ee')], tol=1e-4, verbose=True)
+        frfil.run_tophat_frfilter(to_filter=[(53, 54, 'ee')], tol=1e-4, verbose=True)
         np.testing.assert_almost_equal(d[(53, 54, 'ee')], frfil.clean_resid[(53, 54, 'ee')], decimal=5)
         np.testing.assert_array_equal(f[(53, 54, 'ee')], frfil.flags[(53, 54, 'ee')])
+        # test NotImplementedError
+        pytest.raises(NotImplementedError, frf.load_tophat_frfilter_and_write, uvh5, res_outfilename=outfilename, tol=1e-4,
+                      clobber=True, Nbls_per_load=1, avg_red_bllens=True, baseline_list=[(54, 54, 'ee')])
 
         # test loading and writing all baselines at once.
         uvh5 = os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.OCR_53x_54x_only.uvh5")
         outfilename = os.path.join(tmp_path, 'temp.h5')
         for avg_bl in [True, False]:
-            frf.load_fr_filter_and_write(uvh5, res_outfilename=outfilename, tol=1e-4, clobber=True,
-                                         Nbls_per_load=None, avg_red_bllens=avg_bl)
+            frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename, tol=1e-4, clobber=True,
+                                                Nbls_per_load=None, avg_red_bllens=avg_bl)
             hd = io.HERAData(outfilename)
             d, f, n = hd.read(bls=[(53, 54, 'ee')])
             for bl in d:
@@ -414,7 +418,7 @@ class Test_FRFilter(object):
 
         frfil = frf.FRFilter(uvh5, filetype='uvh5')
         frfil.read(bls=[(53, 54, 'ee')])
-        frfil.run_fr_filter(to_filter=[(53, 54, 'ee')], tol=1e-4, verbose=True)
+        frfil.run_tophat_frfilter(to_filter=[(53, 54, 'ee')], tol=1e-4, verbose=True)
         np.testing.assert_almost_equal(d[(53, 54, 'ee')], frfil.clean_resid[(53, 54, 'ee')], decimal=5)
         np.testing.assert_array_equal(f[(53, 54, 'ee')], frfil.flags[(53, 54, 'ee')])
 
@@ -422,8 +426,8 @@ class Test_FRFilter(object):
         outfilename = os.path.join(tmp_path, 'temp.h5')
         os.remove(outfilename)
         for avg_bl in [True, False]:
-            frf.load_fr_filter_and_write(uvh5, calfile=cal, tol=1e-4, res_outfilename=outfilename,
-                                         Nbls_per_load=2, clobber=True, avg_red_bllens=avg_bl)
+            frf.load_tophat_frfilter_and_write(uvh5, calfile_list=cal, tol=1e-4, res_outfilename=outfilename,
+                                                Nbls_per_load=2, clobber=True, avg_red_bllens=avg_bl)
             hd = io.HERAData(outfilename)
             assert 'Thisfilewasproducedbythefunction' in hd.history.replace('\n', '').replace(' ', '')
             d, f, n = hd.read()
@@ -453,8 +457,8 @@ class Test_FRFilter(object):
         # and entire final channel being flagged
         # when flags are broadcasted.
         time_thresh = 2. / hd.Ntimes
-        frf.load_fr_filter_and_write(input_file, res_outfilename=outfilename, tol=1e-4,
-                                     factorize_flags=True, time_thresh=time_thresh, clobber=True)
+        frf.load_tophat_frfilter_and_write(input_file, res_outfilename=outfilename, tol=1e-4,
+                                            factorize_flags=True, time_thresh=time_thresh, clobber=True)
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for bl in f:
@@ -462,8 +466,8 @@ class Test_FRFilter(object):
             assert np.all(f[bl][0, :])
 
         # test delay filtering and writing with factorized flags and partial i/o
-        frf.load_fr_filter_and_write(input_file, res_outfilename=outfilename, tol=1e-4,
-                                     factorize_flags=True, time_thresh=time_thresh, clobber=True)
+        frf.load_tophat_frfilter_and_write(input_file, res_outfilename=outfilename, tol=1e-4,
+                                            factorize_flags=True, time_thresh=time_thresh, clobber=True)
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for bl in f:
@@ -472,8 +476,8 @@ class Test_FRFilter(object):
             assert np.all(f[bl][:, -1])
             assert not np.all(np.isclose(d[bl], 0.))
 
-        frf.load_fr_filter_and_write(input_file, res_outfilename=outfilename, tol=1e-4, Nbls_per_load=1,
-                                     factorize_flags=True, time_thresh=time_thresh, clobber=True)
+        frf.load_tophat_frfilter_and_write(input_file, res_outfilename=outfilename, tol=1e-4, Nbls_per_load=1,
+                                            factorize_flags=True, time_thresh=time_thresh, clobber=True)
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for bl in f:
@@ -491,60 +495,116 @@ class Test_FRFilter(object):
         uvf.flag_array[:] = False
         flagfile = os.path.join(tmp_path, 'test_flag.h5')
         uvf.write(flagfile, clobber=True)
-        frf.load_fr_filter_and_write(uvh5, res_outfilename=outfilename,
-                                     Nbls_per_load=1, clobber=True, mode='dpss_leastsq',
-                                     external_flags=flagfile,
-                                     overwrite_flags=True)
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename,
+                                            Nbls_per_load=1, clobber=True, mode='dayenu',
+                                            external_flags=flagfile,
+                                            overwrite_flags=True)
         # test that all flags are False
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for k in f:
             assert np.all(~f[k])
         # now without parital io.
-        frf.load_fr_filter_and_write(uvh5, res_outfilename=outfilename,
-                                     clobber=True, mode='dpss_leastsq',
-                                     external_flags=flagfile,
-                                     overwrite_flags=True)
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename,
+                                            clobber=True, mode='dayenu',
+                                            external_flags=flagfile,
+                                            overwrite_flags=True)
         # test that all flags are False
         hd = io.HERAData(outfilename)
         d, f, n = hd.read(bls=[(53, 54, 'ee')])
         for k in f:
             assert np.all(~f[k])
 
-    def test_frf_clean_argparser(self):
-        sys.argv = [sys.argv[0], 'a', '--clobber', '--window', 'blackmanharris']
-        parser = frf.frate_filter_argparser()
+    def test_load_dayenu_filter_and_write(self, tmpdir):
+        tmp_path = tmpdir.strpath
+        uvh5 = os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.OCR_53x_54x_only.uvh5")
+        cdir = os.path.join(tmp_path, 'cache_temp')
+        # make a cache directory
+        if os.path.isdir(cdir):
+            shutil.rmtree(cdir)
+        os.mkdir(cdir)
+        outfilename = os.path.join(tmp_path, 'temp.h5')
+        # run dayenu filter
+        avg_bl = True
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename,
+                                            cache_dir=cdir, mode='dayenu',
+                                            Nbls_per_load=1, clobber=True, avg_red_bllens=avg_bl,
+                                            spw_range=(0, 32), write_cache=True)
+        # generate duplicate cache files to test duplicate key handle for cache load.
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename, cache_dir=cdir,
+                                            mode='dayenu', avg_red_bllens=avg_bl,
+                                            Nbls_per_load=1, clobber=True, read_cache=False,
+                                            spw_range=(0, 32), write_cache=True)
+        # there should now be six cache files (one per i/o/filter). There are three baselines.
+        assert len(glob.glob(cdir + '/*')) == 6
+        hd = io.HERAData(outfilename)
+        assert 'Thisfilewasproducedbythefunction' in hd.history.replace('\n', '').replace(' ', '')
+        d, f, n = hd.read(bls=[(53, 54, 'ee')])
+        np.testing.assert_array_equal(f[(53, 54, 'ee')], True)
+        os.remove(outfilename)
+        shutil.rmtree(cdir)
+        os.mkdir(cdir)
+        # now do all the baselines at once.
+        for avg_bl in [True, False]:
+            frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename,
+                                                cache_dir=cdir, mode='dayenu', avg_red_bllens=avg_bl,
+                                                Nbls_per_load=None, clobber=True,
+                                                spw_range=(0, 32), write_cache=True)
+            if avg_bl:
+                assert len(glob.glob(cdir + '/*')) == 1
+            hd = io.HERAData(outfilename)
+            assert 'Thisfilewasproducedbythefunction' in hd.history.replace('\n', '').replace(' ', '')
+            d, f, n = hd.read(bls=[(53, 54, 'ee')])
+            np.testing.assert_array_equal(f[(53, 54, 'ee')], True)
+            os.remove(outfilename)
+        shutil.rmtree(cdir)
+        os.mkdir(cdir)
+        # run again using computed cache.
+        calfile = os.path.join(DATA_PATH, "test_input/zen.2458101.46106.xx.HH.uv.abs.calfits_54x_only")
+        frf.load_tophat_frfilter_and_write(uvh5, res_outfilename=outfilename, max_frate_coeffs=[0.0, 0.025],
+                                            cache_dir=cdir, calfile_list=calfile, read_cache=True,
+                                            Nbls_per_load=1, clobber=True, mode='dayenu',
+                                            spw_range=(0, 32), write_cache=True)
+        # no new cache files should be generated.
+        assert len(glob.glob(cdir + '/*')) == 1
+        hd = io.HERAData(outfilename)
+        assert 'Thisfilewasproducedbythefunction' in hd.history.replace('\n', '').replace(' ', '')
+        d, f, n = hd.read(bls=[(53, 54, 'ee')])
+        np.testing.assert_array_equal(f[(53, 54, 'ee')], True)
+        os.remove(outfilename)
+        shutil.rmtree(cdir)
+
+    def test_tophat_clean_argparser(self):
+        sys.argv = [sys.argv[0], 'a', '--clobber', '--window', 'blackmanharris', '--max_frate_coeffs', '0.024', '-0.229']
+        parser = frf.tophat_frfilter_argparser()
         a = parser.parse_args()
-        assert a.infilename == 'a'
+        assert a.datafilelist == ['a']
         assert a.clobber is True
         assert a.window == 'blackmanharris'
+        assert a.max_frate_coeffs[0] == 0.024
+        assert a.max_frate_coeffs[1] == -0.229
         assert a.time_thresh == 0.05
         assert not a.factorize_flags
-        assert a.frate_standoff == 0.0
-        assert a.frac_frate_sky_max == 1.0
-        assert a.min_frate == 0.025
 
-    def test_frf_linear_argparser(self):
-        sys.argv = [sys.argv[0], 'a', '--clobber', '--write_cache', '--cache_dir', '/blah/', '--min_frate', '0.1', '--frate_standoff', '3.2', '--frac_frate_sky_max', '.99']
-        parser = frf.frate_filter_argparser(mode='dayenu')
+    def test_tophat_linear_argparser(self):
+        sys.argv = [sys.argv[0], 'a', '--clobber', '--write_cache', '--cache_dir', '/blah/', '--max_frate_coeffs', '0.024', '-0.229']
+        parser = frf.tophat_frfilter_argparser(mode='dayenu')
         a = parser.parse_args()
-        assert a.infilename == 'a'
+        assert a.datafilelist == ['a']
         assert a.clobber is True
         assert a.write_cache is True
         assert a.cache_dir == '/blah/'
+        assert a.max_frate_coeffs[0] == 0.024
+        assert a.max_frate_coeffs[1] == -0.229
         assert a.time_thresh == 0.05
-        assert a.frate_standoff == 3.2
-        assert a.frac_frate_sky_max == .99
-        assert a.min_frate == 0.1
         assert not a.factorize_flags
-        parser = frf.frate_filter_argparser(mode='dpss_leastsq')
+        parser = frf.tophat_frfilter_argparser(mode='dpss_leastsq')
         a = parser.parse_args()
-        assert a.infilename == 'a'
+        assert a.datafilelist == ['a']
         assert a.clobber is True
         assert a.write_cache is True
         assert a.cache_dir == '/blah/'
+        assert a.max_frate_coeffs[0] == 0.024
+        assert a.max_frate_coeffs[1] == -0.229
         assert a.time_thresh == 0.05
         assert not a.factorize_flags
-        assert a.frate_standoff == 3.2
-        assert a.frac_frate_sky_max == .99
-        assert a.min_frate == 0.1
